barman keep: Delete by backup ID in barman-cloud
Updates the implementation of `barman-cloud-delete-backup` so that it
will not delete an archival backup, i.e. a backup with a keep
annotation.

This changes the behaviour of the `barman-cloud-backup-delete` command
such that a user cannot delete a backup which has a keep annotation -
instead the command fails and they get a helpful error message.

This is part of the implementation for #370.

+++ barman/clients/cloud_backup_delete.py
@@ -205,6 +205,18 @@ def main(args=None):
                 raise SystemExit(1)
 
             if config.backup_id:
+                # Because we only care about one backup, skip the annotation cache
+                # because it is only helpful when dealing with multiple backups
+                if catalog.should_keep_backup(config.backup_id, use_cache=False):
+                    logging.error(
+                        "Skipping delete of backup %s for server %s "
+                        "as it has a current keep request. If you really "
+                        "want to delete this backup please remove the keep "
+                        "and try again.",
+                        config.backup_id,
+                        config.server_name,
+                    )
+                    raise SystemExit(1)
                 _delete_backup(
                     cloud_interface, catalog, config.backup_id, config.dry_run
                 )
